+++ packages/governance/src/timelock/mod.rs
@@ -93,11 +93,11 @@ pub const TIMELOCK_EXTEND_AMOUNT: u32 = 30 * DAY_IN_LEDGERS;
 pub const TIMELOCK_TTL_THRESHOLD: u32 = TIMELOCK_EXTEND_AMOUNT - DAY_IN_LEDGERS;
 
 /// Sentinel value for an operation that has not been scheduled.
-pub const UNSET_TIMESTAMP: u32 = 0;
+pub const UNSET_TIMESTAMP: u64 = 0;
 
 /// Sentinel value used to mark an operation as done.
 /// Using 1 instead of 0 to distinguish from unset operations.
-pub const DONE_TIMESTAMP: u32 = 1;
+pub const DONE_TIMESTAMP: u64 = 1;
 
 // ################## EVENTS ##################
 
@@ -146,7 +146,7 @@ pub struct OperationScheduled {
 /// * `args` - The arguments to pass to the function.
 /// * `predecessor` - The predecessor operation ID.
 /// * `salt` - The salt for uniqueness.
-/// * `delay` - The delay in ledgers.
+/// * `delay` - The delay in seconds.
 #[allow(clippy::too_many_arguments)]
 pub fn emit_operation_scheduled(
     e: &Env,
